widnode-supervisor/userapp/requirements.txt
bleak==0.22.2
httpx[http2]==0.27.2
python-dotenv==1.0.1
numpy==1.26.4
dbus-next==0.2.3
psutil==5.9.8
pymodbus==3.6.9

